Changed too much to fit in the brief:
* the chapter of l0 was completely relocated and renamed
It is now part of Terminologie & Kontext
as it actually provides just that.
* Updated the content of this previous chapter l0 and updated my math.
* Tick has been relocated to sit exactly after the definition of time
* added a small subsection to describe which parts of this where the
status quo at the beginning of the project and which parts where newly
added. This section should be moved further down later, but I will stop
refactoring here. Untouched therefore are chapters about Hitbox and
down from there.